requirements.txt
alembic==1.7.7
aniso8601==9.0.1
attrs==21.4.0
black==22.3.0
boto3==1.24.51
botocore==1.27.51
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.2
Flask==2.1.1
Flask-Migrate==3.1.0
flask-restx==0.5.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
jmespath==1.0.1
jsonschema==4.4.0
Mako==1.2.0
MarkupSafe==2.1.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.2
psycopg-binary==3.0.11
psycopg2-binary==2.9.3
pyrsistent==0.18.1
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
requests==2.28.1
s3transfer==0.6.0
six==1.16.0
SQLAlchemy==1.4.35
tomli==2.0.1
typing-extensions==4.2.0
urllib3==1.26.11
Werkzeug==2.1.1
zipp==3.8.0
